Update README.md
The reason for this change is that in rails4 the "<%= include_gon %>" throws 'gon not found' error. 
I was facing this issue with my rails 4.1 app. And it wasted my time a lot. I would suggest that for now, if we can add the solution given in #90 in the readme file will help to others a lot.