Stop using Travis CI and switch to GitHub Actions
We previously used Travis CI conveniently, but as of 2025, [using Travis CI to build OSS requires contacting them to get a fixed amount of credits each month](https://docs.travis-ci.com/user/billing-faq/#what-if-i-am-building-open-source), making it cumbersome. I want to reduce the hassle by migrating from Travis CI to GitHub Actions.

In the migration to GitHub Actions, we have configured CI to work with Ruby up to version 3.4.

The CI will pass even if the tests fail on `head`.
Currently, the tests fail on `head`.
It should pass after merging #273.

The CI will also pass even if the tests fail on `truffleruby`.
One test is currently failing.
I’d like to leave the investigation of the cause to someone else.

Also made the following changes to make the tests pass.

- Remove the version lock on `rabl` because it was causing test failures.
- Stop using `File.exists?`, which was removed in Ruby 3.2, and switched to `File.exist?` instead.
- Replace `controller.view_paths << 'spec/test_data'` with `append_view_path` because it caused a `NoMethodError`.
- Avoided the `NoMethodError` on [this line](https://github.com/gazay/gon/blob/2d6f920393dedbe1af6a5be5c9317cd54b2c8357/lib/gon/jbuilder/parser.rb#L48) by using a stub.
- Locked `concurrent-ruby` to version 1.3.4 because using version 1.3.5 or higher with the old Rails caused errors
  - ref: [Rails 7.0.8 fails to create an app with most recent concurrent-ruby version · Issue #54260 · rails/rails](https://github.com/rails/rails/issues/54260)
- Lock the version of `loofah` when using Ruby versions below 2.5 because dependency issues with `nokogiri` caused test failures on Ruby 2.4 and Rails 5.2 or lower.