Remove ostruct dependency
ostruct has become a bundled gem, so the following warning is output
from Ruby 3.4:

```
/path/to/ruby/gems/3.4.0/gems/gon-6.4.0/lib/gon/base.rb:1:
warning: ostruct was loaded from the standard library, but will no
longer be part of the default gems starting from Ruby 3.5.0. You can
add ostruct to your Gemfile or gemspec to silence this warning.
```

We will change to use Struct instead and remove the dependency.