Fixed a test that sometimes failed
The test was removing a Rails constant, and when another test tried to call Rails.version afterward, it caused failures. Since there’s no need to remove the constant, I stopped doing so to fix the issue.

```
  1) threading behaviour is threadsafe
     Failure/Error: skip 'ActionDispatch::TestRequest.create is not supported on Rails versions below 5.0' if Rails.version < '5.0'

     NameError:
       uninitialized constant Rails
     # ./spec/gon/thread_spec.rb:24:in 'block (2 levels) in <top (required)>'
```
The test was removing a Rails constant, and when another test tried to call Rails.version afterward, it caused failures. Since there’s no need to remove the constant, I stopped doing so to fix the issue.